pyproject.toml
[project]
name = "ai-agents-email-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=1.1.1",
    "minsearch>=0.0.10",
    "numpy>=2.4.3",
    "pydantic-ai>=1.72.0",
    "python-dotenv>=1.2.2",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
